requirements.txt
bcrypt==3.1.4
black==22.6.0
blinker==1.4
cffi==1.14.2
click==8.1.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==1.1.2
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.1.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.8.4
pycparser==2.19
six==1.11.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing-extensions==4.3.0
Werkzeug==0.14.1
WTForms==2.2.1
